requirements.txt
requests==2.27.1
BeautifulSoup4~=4.11.1
schema==0.7.5
tqdm==4.64.0
SQLAlchemy~=1.4.41
fastapi==0.85.0
uvicorn==0.18.3
Faker==13.12.1
pymysql==1.0.2
pydantic~=1.10.2
python-multipart==0.0.5
Pillow==9.2.0
